# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Capture API Simulator API",
    "description": "Deterministic traffic-forensics simulator. Browsers cannot call it directly (no CORS): proxy it from your own server. Optional fields are absent rather than null; times are ISO-8601 UTC with milliseconds; session ids are uint64 decimal strings.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/enrich/ips": {
      "post": {
        "tags": [
          "enrichment"
        ],
        "summary": "Enrich up to 100 addresses at once",
        "description": "Looks up every address in one call and answers a map keyed by the address exactly as it was sent. Internal and unparseable addresses come back with `reputation: \"unknown\"` and no country, ASN or org (absent keys, never null). Limited to 1 request per second per session family, so coalesce a grid page into a single call and cache the answers.",
        "operationId": "enrichIps",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrichRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrichResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Semantic validation error",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorBody"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Enrichips"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; honour Retry-After",
            "headers": {
              "Retry-After": {
                "description": "Integer seconds (an HTTP-date on /v1/auth/login).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/enrich/ips/{ip}": {
      "get": {
        "tags": [
          "enrichment"
        ],
        "summary": "Enrich one address",
        "description": "The same object the batch route returns for a single address. Limited to 5 requests per second per session family; use `POST /v1/enrich/ips` for anything bigger than a detail pane.",
        "operationId": "getIpEnrichment",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "ip",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 45,
              "description": "IPv4 or IPv6 address.",
              "title": "Ip"
            },
            "description": "IPv4 or IPv6 address."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrichResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; honour Retry-After",
            "headers": {
              "Retry-After": {
                "description": "Integer seconds (an HTTP-date on /v1/auth/login).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/detections": {
      "get": {
        "tags": [
          "detections"
        ],
        "summary": "Detections released so far",
        "description": "Ascending by `seq`. Without `after_seq` the newest `limit` detections are returned (still ascending). `last_seq` is the head of the whole feed, so polling with `after_seq=last_seq` never re-scans detections on sensors the caller cannot read.",
        "operationId": "listDetections",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "after_seq",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return seq greater than this.",
              "title": "After Seq"
            },
            "description": "Return seq greater than this."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "sensor_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated sensor ids; defaults to every sensor the caller may read.",
              "examples": [
                "hq-core,harbor-branch"
              ],
              "title": "Sensor Ids"
            },
            "description": "Comma-separated sensor ids; defaults to every sensor the caller may read."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionList"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Semantic validation error",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorBody"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorBody"
                    }
                  ],
                  "title": "Response 422 Listdetections"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stream/detections": {
      "get": {
        "tags": [
          "detections"
        ],
        "summary": "Live detection feed (Server-Sent Events)",
        "description": "Opens with a `: open last_seq=<n>` comment and `retry: 3000`. A resume point (`Last-Event-ID` header, which wins, or `?last_event_id`) replays the ring of the last 1,000 detections; a resume point older than the ring yields a `reset` event instead. Live events carry `id: <seq>` and `event: detection`. The stream ends with a `reauth` event when the opening access token expires, silently when the token family is revoked, and on server rotation (300 s calm, 60 s flaky, 20 s storm).",
        "operationId": "streamDetections",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "name": "last_event_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Resume point; the Last-Event-ID header wins.",
              "title": "Last Event Id"
            },
            "description": "Resume point; the Last-Event-ID header wins."
          },
          {
            "name": "sensor_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated sensor ids; defaults to every sensor the caller may read.",
              "examples": [
                "hq-core,harbor-branch"
              ],
              "title": "Sensor Ids"
            },
            "description": "Comma-separated sensor ids; defaults to every sensor the caller may read."
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, invalid or revoked credentials",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer challenge (Bearer-authenticated routes).",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Detection": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts",
            "examples": [
              "2025-10-27T09:14:03.120Z"
            ]
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "mitre": {
            "$ref": "#/components/schemas/Mitre"
          },
          "sensor_id": {
            "type": "string",
            "title": "Sensor Id"
          },
          "session_id": {
            "type": "string",
            "pattern": "^[0-9]{1,20}$",
            "title": "Session Id",
            "description": "uint64 as a decimal string; larger than 2**53, never parse as a number.",
            "examples": [
              "72075232438042624"
            ]
          },
          "src": {
            "$ref": "#/components/schemas/Endpoint"
          },
          "dst": {
            "$ref": "#/components/schemas/Endpoint"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "seq",
          "id",
          "ts",
          "rule_id",
          "rule",
          "severity",
          "mitre",
          "sensor_id",
          "session_id",
          "src",
          "dst",
          "summary"
        ],
        "title": "Detection"
      },
      "DetectionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Detection"
            },
            "type": "array",
            "title": "Items",
            "description": "Ascending seq."
          },
          "last_seq": {
            "type": "integer",
            "title": "Last Seq"
          }
        },
        "type": "object",
        "required": [
          "items",
          "last_seq"
        ],
        "title": "DetectionList"
      },
      "Endpoint": {
        "properties": {
          "ip": {
            "type": "string",
            "title": "Ip"
          },
          "port": {
            "type": "integer",
            "title": "Port"
          },
          "host": {
            "type": "string",
            "title": "Host"
          },
          "country": {
            "type": "string",
            "title": "Country",
            "description": "ISO 3166-1 alpha-2; absent internally."
          }
        },
        "type": "object",
        "required": [
          "ip",
          "port"
        ],
        "title": "Endpoint"
      },
      "EnrichRequest": {
        "properties": {
          "ips": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Ips"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ips"
        ],
        "title": "EnrichRequest"
      },
      "EnrichResponse": {
        "properties": {
          "results": {
            "additionalProperties": {
              "$ref": "#/components/schemas/EnrichResult"
            },
            "type": "object",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "EnrichResponse"
      },
      "EnrichResult": {
        "properties": {
          "country": {
            "type": "string",
            "title": "Country"
          },
          "asn": {
            "type": "string",
            "title": "Asn"
          },
          "org": {
            "type": "string",
            "title": "Org"
          },
          "reputation": {
            "$ref": "#/components/schemas/Reputation"
          }
        },
        "type": "object",
        "required": [
          "reputation"
        ],
        "title": "EnrichResult"
      },
...
      "ErrorBody": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable message."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable snake_case error code."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorBody",
        "description": "Error envelope of every domain error. Extra keys carry machine-readable context."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Mitre": {
        "properties": {
          "technique_id": {
            "type": "string",
            "title": "Technique Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "technique_id",
          "name"
        ],
        "title": "Mitre"
      },
...
      "Reputation": {
        "type": "string",
        "enum": [
          "clean",
          "suspicious",
          "malicious",
          "unknown"
        ]
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationErrorBody": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ValidationErrorBody",
        "description": "FastAPI's default 422 body: detail is a LIST here, unlike ErrorBody."
      },
      "ValidationIssue": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "ctx": {
            "additionalProperties": true,
            "type": "object",
            "title": "Ctx"
          },
          "input": {
            "title": "Input"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationIssue",
        "description": "One entry of FastAPI's default request-validation error list."
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Access token `at_…` from POST /v1/auth/login or /v1/auth/refresh.",
        "scheme": "bearer"
      }
    }
  },
...
}